project/first/requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
gitpython>=3.1.30
python-dotenv>=1.0.0

# API dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0

# OpenAI API
openai>=1.5.0

# Utils
requests>=2.31.0
